test: reorganize backup tests into correct feature and unit test directories
